Disabled the ability to edit the names of existing files to address an issue where changing the name could fragment the file chain.
Renaming the existing files is not a practical solution as the file storage system API in Django does not provide a rename method (or a similar mechanism). An alternate approach, should this feature be deemed necessary, is to name files based solely on the primary key of the key (see the upload_path function in bdr.utils.storage) which should be immutable and independent of its display name. However, this would be a backwards-incompatible change requiring direct modification of the database as well as moving all existing files.

bdr.models.base:
 # Marked the name field of the File class as uneditable.